requirements.txt
streamlit==1.50.0
pandas==2.3.3
plotly==6.3.0
requests==2.32.5
websocket-client==1.8.0
